Move reducers/add_calls to reducers/index.
This is more in line with the pattern for the rest of this project.
